refactor(namespaces for services) Creating namespaces for services
- Creating namespace for Accounts
- Creating namespace for Categories
- Creating namespace for Transactions
- Creating namespace for Goals
- Updating all calls to services with namespaces
