+++ app/controllers/goals_controller.rb
@@ -13,11 +13,11 @@ class GoalsController < ApplicationController
   def new
     @goal = Goal.new
 
-    @current_balance = UpdateGoalsCurrentBalanceService.new(current_user).calculate_current_balance
+    @current_balance = Goals::UpdateGoalsCurrentBalanceService.new(current_user).calculate_current_balance
   end
 
   def edit
-    @current_balance = UpdateGoalsCurrentBalanceService.new(current_user).calculate_current_balance
+    @current_balance = Goals::UpdateGoalsCurrentBalanceService.new(current_user).calculate_current_balance
   end
 
   def create
