--- app/services/account_search_service.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-class AccountSearchService
-  def initialize(user)
-    @user = user
-  end
-
-  def call
-    @user.accounts
-  end
-end
