--- app/services/calculate_total_expense.rb
@@ -1,17 +1,0 @@
-# frozen_string_literal: true
-
-class CalculateTotalExpense
-  def initialize(account_id)
-    @account_id = account_id
-  end
-
-  def call
-    calculate_expenses
-  end
-
-  private
-
-  def calculate_expenses
-    Transaction.where(account_id: @account_id, transaction_type: "expense", status: "completed").sum(:amount)
-  end
-end
