+++ app/services/goals/update_goals_current_balance_service.rb
@@ -1,0 +1,20 @@
+# frozen_string_literal: true
+
+module Goals
+  class UpdateGoalsCurrentBalanceService
+    def initialize(user)
+      @user = user
+    end
+
+    def call
+      total_balance = @user.total_account_balance
+      @user.goals.find_each do |goal|
+        goal.update(current_amount: total_balance)
+      end
+    end
+
+    def calculate_current_balance
+      @user.total_account_balance
+    end
+  end
+end
